Prep for multiple neural networks
